refactor: rename path home for Home
